refactor(editor): rename useFieldValueStore, and write down what it captures
I claimed field values could only ever exist for logs streams, and repeated it
after being questioned. Wrong twice over: the sidebar path captureFromValuesApi
is reached from logs, traces AND pipeline, and captureFromSearchHits captures
under whichever stream type the Logs page searched -- its selector covers
metrics and traces. Proved by searching cache_hit_ratio in the Logs UI, after
which the store held 21 metrics keys and the SLO form resolved environment ->
[development, staging].

Two names made that easy to get wrong, so both are fixed rather than just
apologised for:

  useFieldValueStore.ts -> fieldValueStore.ts

It is not a composable -- plain functions over IndexedDB, no reactive state --
and the false `use` prefix left it one character from useFieldValuesStream,
which IS a composable and does something else entirely. Dropping the prefix
also puts it next to its sibling fieldValueDB.ts, where it belongs.

The header now names BOTH writers, states that neither is tied to a stream type,
and spells out the trap directly: composables/useLogs/ is the Logs PAGE, not the
logs stream TYPE. StreamContext.streamType says the same at the field level --
part of the key, never a filter on what may be captured.

The rule worth remembering: values come from what has been SEARCHED or EXPANDED.
A stream nobody has looked at has none. That is a cold cache, not a missing
capability -- which is exactly the distinction I failed to make, off one grep.

Mechanical rename across 10 files including the vi.mock paths. 140 passing in
the touched specs; type-check, prettier clean; the 13 eslint warnings in
IndexList.vue are pre-existing (identical count with this change stashed).
